Created a sign-up form linked with firebase, missing routing
