+++ src/pages/Firebase/firebase.js
@@ -16,6 +16,8 @@ class Firebase {
     this.auth = app.auth();
   }
 
+  // Autherisation API
+  
   doCreateUserWithEmailAndPassword = (email, password) =>
     this.auth.createUserWithEmailAndPassword(email, password);
 
@@ -25,7 +27,7 @@ class Firebase {
   doSignOut = () => this.auth.signOut();
 
   doPasswordReset = email => this.auth.sendPasswordResetEmail(email);
-  
+
   doPasswordUpdate = password =>
     this.auth.currentUser.updatePassword(password);
 }
