+++ src/pages/Firebase/index.js
@@ -1,6 +1,6 @@
-import FirebaseContext from './context';
+import FirebaseContext, { withFirebase } from './context';
 import Firebase from './firebase';
 
 export default Firebase
 
-export { FirebaseContext };
+export { FirebaseContext, withFirebase };
