+++ src/pages/Navigation/index.js
@@ -10,6 +10,9 @@ const Naviation = () => (
         <Link to={ROUTES.SIGN_IN}>Sign In</Link>
       </li>
       <li>
+        <Link to={ROUTES.SIGN_UP}>Sign Up</Link>
+      </li>
+      <li>
         <Link to={ROUTES.LANDING}>Landing</Link>
       </li>
       <li>
